requirements.txt
fastapi
strawberry-graphql
uvicorn
motor
python-dotenv
openai
fpdf2
python-jose[cryptography]
httpx
